tests/isolation/requirements.txt
pytest>=8.0
supabase>=2.7
pyjwt>=2.9
